+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- SnapshotOptions can be called as a Expression instead of passing the SnapshotOptions object directly
 
 ## v2.5.0
 ### Added