+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.5.1
+### Added
+- SnapshotOptions can be called as a Expression instead of passing the SnapshotOptions object directly
+- Clone for SnapshotOptions
+- GitLab Pipeline
+
 ## v2.5.0
 ### Added
 - Allow the SnapshotTokenizer to accept a IObjectMapper that has to be used